chore: improve InputGroup exports
